Fix voltage and pressure readings to match Entacore's software
With the Altimeter infrastructure in place we can apply the ADC offset.
In addition, I was miscalculating the pressure reading based on 5.1VDC
instead of 5.0.  Even if it were 5.1, the ADC's basis then also changes
to 5.1 and the result should be roughly the same.

In any case, the ADC, Voltage, and Pressure readings all match
Entacore's output.  The Altitude is currently in feet and I haven't
checked it yet.